+++ packages/cdn/angular/mui-angular.js
@@ -681,8 +681,9 @@ function enableScrollLockFn() {
 
 /**
  * Turn off window scroll lock.
+ * @param {Boolean} resetPos - Reset scroll position to original value.
  */
-function disableScrollLockFn() {
+function disableScrollLockFn(resetPos) {
   // ignore
   if (scrollLock === 0) return;
 
@@ -695,7 +696,7 @@ function disableScrollLockFn() {
         doc = document;
 
     jqLite.removeClass(doc.body, scrollLockCls);
-    win.scrollTo(scrollLockPos.left, scrollLockPos.top);
+    if (resetPos) win.scrollTo(scrollLockPos.left, scrollLockPos.top);
   }
 }
 
@@ -1778,7 +1779,7 @@ _angular2.default.module(moduleName, []).directive('muiSelect', ['$timeout', fun
           selectEl[0].focus();
 
           // disable scroll lock
-          util.disableScrollLock();
+          util.disableScrollLock(true);
 
           // remove event handlers
           jqLite.off(document, 'click', closeMenuFn);