+++ packages/cdn/react/mui-react.js
@@ -775,8 +775,9 @@ function enableScrollLockFn() {
 
 /**
  * Turn off window scroll lock.
+ * @param {Boolean} resetPos - Reset scroll position to original value.
  */
-function disableScrollLockFn() {
+function disableScrollLockFn(resetPos) {
   // ignore
   if (scrollLock === 0) return;
 
@@ -789,7 +790,7 @@ function disableScrollLockFn() {
         doc = document;
 
     jqLite.removeClass(doc.body, scrollLockCls);
-    win.scrollTo(scrollLockPos.left, scrollLockPos.top);
+    if (resetPos) win.scrollTo(scrollLockPos.left, scrollLockPos.top);
   }
 }
 
@@ -2911,7 +2912,7 @@ var Select = function (_React$Component) {
     key: 'hideMenu',
     value: function hideMenu() {
       // remove scroll lock
-      util.disableScrollLock();
+      util.disableScrollLock(true);
 
       // remove event listeners
       jqLite.off(window, 'resize', this.hideMenuCB);
@@ -2953,11 +2954,10 @@ var Select = function (_React$Component) {
       var className = _props.className;
       var style = _props.style;
       var label = _props.label;
-      var value = _props.value;
       var defaultValue = _props.defaultValue;
       var readOnly = _props.readOnly;
       var useDefault = _props.useDefault;
-      var reactProps = babelHelpers.objectWithoutProperties(_props, ['children', 'className', 'style', 'label', 'value', 'defaultValue', 'readOnly', 'useDefault']);
+      var reactProps = babelHelpers.objectWithoutProperties(_props, ['children', 'className', 'style', 'label', 'defaultValue', 'readOnly', 'useDefault']);
 
 
       return _react2.default.createElement(
@@ -2973,7 +2973,7 @@ var Select = function (_React$Component) {
           'select',
           babelHelpers.extends({}, reactProps, {
             ref: 'selectEl',
-            value: value,
+            value: this.state.value,
             defaultValue: defaultValue,
             readOnly: this.props.readOnly,
             onChange: this.onInnerChangeCB,