+++ packages/npm/lib/react/select.js
@@ -196,7 +196,7 @@ var Select = function (_React$Component) {
     key: 'hideMenu',
     value: function hideMenu() {
       // remove scroll lock
-      util.disableScrollLock();
+      util.disableScrollLock(true);
 
       // remove event listeners
       jqLite.off(window, 'resize', this.hideMenuCB);
@@ -238,11 +238,10 @@ var Select = function (_React$Component) {
       var className = _props.className;
       var style = _props.style;
       var label = _props.label;
-      var value = _props.value;
       var defaultValue = _props.defaultValue;
       var readOnly = _props.readOnly;
       var useDefault = _props.useDefault;
-      var reactProps = babelHelpers.objectWithoutProperties(_props, ['children', 'className', 'style', 'label', 'value', 'defaultValue', 'readOnly', 'useDefault']);
+      var reactProps = babelHelpers.objectWithoutProperties(_props, ['children', 'className', 'style', 'label', 'defaultValue', 'readOnly', 'useDefault']);
 
 
       return _react2.default.createElement(
@@ -258,7 +257,7 @@ var Select = function (_React$Component) {
           'select',
           babelHelpers.extends({}, reactProps, {
             ref: 'selectEl',
-            value: value,
+            value: this.state.value,
             defaultValue: defaultValue,
             readOnly: this.props.readOnly,
             onChange: this.onInnerChangeCB,